+++ dpnp/scipy/linalg/_decomp_lu.py
@@ -42,7 +42,10 @@ It contains:
 
 
 import dpnp
-from dpnp.linalg.dpnp_utils_linalg import assert_stacked_2d
+from dpnp.linalg.dpnp_utils_linalg import (
+    assert_stacked_2d,
+    assert_stacked_square,
+)
 
 from ._utils import (
     dpnp_lu_factor,
@@ -132,7 +135,7 @@ def lu_solve(lu_and_piv, b, trans=0, overwrite_b=False, check_finite=True):
     Parameters
     ----------
     lu, piv : {tuple of dpnp.ndarrays or usm_ndarrays}
-        LU factorization of matrix `a` (M, M) together with pivot indices.
+        LU factorization of matrix `a` (..., M, M) together with pivot indices.
     b : {(M,), (..., M, K)} {dpnp.ndarray, usm_ndarray}
         Right-hand side
     trans : {0, 1, 2} , optional
@@ -160,7 +163,7 @@ def lu_solve(lu_and_piv, b, trans=0, overwrite_b=False, check_finite=True):
 
     Returns
     -------
-    x : {(M,), (M, K)} dpnp.ndarray
+    x : {(M,), (..., M, K)} dpnp.ndarray
         Solution to the system
 
     Warning
@@ -187,6 +190,7 @@ def lu_solve(lu_and_piv, b, trans=0, overwrite_b=False, check_finite=True):
     (lu, piv) = lu_and_piv
     dpnp.check_supported_arrays_type(lu, piv, b)
     assert_stacked_2d(lu)
+    assert_stacked_square(lu)
 
     return dpnp_lu_solve(
         lu,